requirements.txt
# Django & DRF
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Data collection — Reddit API
praw==7.7.1

# Data collection — Web scraping
playwright==1.41.0
beautifulsoup4==4.12.2
requests==2.31.0

# NLP — HuggingFace Transformers (local models)
transformers==4.37.0
torch==2.1.2
sentencepiece==0.1.99

# NLP — spaCy for NER
spacy==3.7.2

# NLP — Language detection
langdetect==1.0.9

# Geo — OpenStreetMap, geocoding, spatial
osmnx==1.9.1
geopy==2.4.1
shapely==2.0.2

# Data processing
pandas==2.1.4
numpy==1.26.3

# Database — Supabase (PostgreSQL)
psycopg2-binary==2.9.6

# Server
gunicorn==20.1.0

# Task queue
redis==5.0.1
celery==5.3.6

# Utilities
python-dotenv==1.0.0
joblib==1.3.2
